converted routes to controllers
